requirements.txt
Flask==2.3.2
#Flask-Bcrypt==0.7.1
Flask-Cors==4.0.0
Flask-Migrate==4.0.4
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.5
PyMySQL==1.0.2
pymssql==2.2.7

MarkupSafe==2.1.3
marshmallow==3.20.1

Werkzeug==2.3.6


